Added tests for CUDA operators
The method of handling CUDA functions from the extension should change. For now we have placeholder functions in the main module that are solely meant to be overloaded. This should probably changed to have CUDA overload the regular CPU variants. Also for some reason being in the test environment does not let the extension load, meaning the CUDA tests aren't running.